FirstClassTestRunner에서 상속 대신 구성 사용
ParentRunner<T> 상속(is-a)을 버리고 구현체를 소유하는 구성(has-a)으로
변경한다. 이 변경은 상속으로 불필요하게 노출되는 BParentRunner<T>
공용멤버를 숨겨 불필요한 결합을 방지한다.

변경과정에서 FirstClassTestCaseMethod는 상위 클래스로 존재할 필요가 없어
inner class로 수정되었다.
